requirements.txt
python-telegram-bot==20.7
tweepy
requests
python-dotenv
Flask
